Updated to the newest spreadsheetlight lib, using doubles instead of floats for monitor counters

+++ vApus.Monitor/MonitorView.cs
@@ -1299,7 +1299,7 @@ namespace vApus.Monitor {
             return monitorControl.MonitorResultCache;
         }
 
-        public Dictionary<DateTime, float[]> GetMonitorValues() {
+        public Dictionary<DateTime, double[]> GetMonitorValues() {
             return monitorControl.GetMonitorValues();
         }
