+++ vApus.Results/Metrics/MonitorMetrics.cs
@@ -25,13 +25,13 @@ namespace vApus.Results {
         public int Run { get; set; }
 
         public string[] Headers { get; set; }
-        public float[] AverageMonitorResults { get; set; }
+        public double[] AverageMonitorResults { get; set; }
         #endregion
 
         #region Constructor
         public MonitorMetrics() {
             Headers = new string[0];
-            AverageMonitorResults = new float[0];
+            AverageMonitorResults = new double[0];
         }
         #endregion
     }
